Refactor employee salary data type to `decimal` and enhance query parameter handling:
- Update `Employee`, `EmployeeDto`, and related services to use `decimal` for Salary.
- Introduce `QueryParamsDto` for consolidating common query parameters.
- Refactor `EmployeeQueryParamsDto` to include `QueryParamsDto`.
- Adjust `EmployeeService` to handle updated query parameter structure.

+++ DTOs/EmployeeDto.cs
@@ -11,7 +11,7 @@ public class EmployeeDto
     public required string FullName { get; set; }
 
     [Range(0, 10000000, ErrorMessage = "Salary must be between 0 and 10000000")]
-    public int Salary { get; set; }
+    public decimal Salary { get; set; }
 
     [Required(ErrorMessage = "IdProfile is required")]
     public int IdProfile { get; set; }
